migrate: refs unit tests to integration tests using TestRepo
